requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
crispy-bootstrap5==2024.10
cssselect2==0.7.0
diff-match-patch==20241021
Django==5.1.4
django-appconf==1.1.0
django-ckeditor==6.7.2
django-crispy-forms==2.3
django-import-export==4.3.3
django-js-asset==3.1.2
django-select2==8.3.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
fonttools==4.56.0
idna==3.10
mysqlclient==2.2.7
numpy==2.2.3
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-leetcode==1.2.1
pytz==2024.2
reportlab==4.3.1
requests==2.32.3
setuptools==77.0.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tablib==3.7.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2024.2
urllib3==2.3.0
weasyprint==64.1
webencodings==0.5.1
wheel==0.45.1
zopfli==0.2.3.post1
